[#11133] Refactor cache IT and remove jobs catalog-common test dep
Move custom catalog test doubles into IcebergCacheIT as inner classes.
Drop maintenance/jobs test dependency on catalog-common by using the
table-metadata-cache-impl property key directly in tests.
