--- iceberg/iceberg-rest-server/src/test/java/org/apache/gravitino/iceberg/integration/test/CustomNoSupportLocationCatalog.java
@@ -1,47 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *  http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package org.apache.gravitino.iceberg.integration.test;
-
-import java.util.List;
-import org.apache.iceberg.Table;
-import org.apache.iceberg.catalog.Catalog;
-import org.apache.iceberg.catalog.Namespace;
-import org.apache.iceberg.catalog.TableIdentifier;
-
-/** Custom catalog that does not support metadata location for cache IT. */
-public class CustomNoSupportLocationCatalog implements Catalog {
-
-  @Override
-  public List<TableIdentifier> listTables(Namespace namespace) {
-    return List.of();
-  }
-
-  @Override
-  public boolean dropTable(TableIdentifier identifier, boolean purge) {
-    return false;
-  }
-
-  @Override
-  public void renameTable(TableIdentifier from, TableIdentifier to) {}
-
-  @Override
-  public Table loadTable(TableIdentifier identifier) {
-    return null;
-  }
-}
